+++ src/main/types.ts
@@ -46,7 +46,7 @@ type UpgradeStatus = {
 };
 
 /**
- * We display any OBS crashes on the frontend so we don't silently recover 
+ * We display any OBS crashes on the frontend so we don't silently recover
  * and have the user think all is well.
  */
 type Crashes = CrashData[];
@@ -174,22 +174,6 @@ type VideoQueueItem = {
   deleteSource: boolean;
   metadata: Metadata;
 };
-
-interface IEventTarget {
-  name: string;
-  value: any;
-}
-
-/**
- * Class to fake an event for onChange in `ISettingsPanelProps`
- */
-class FakeChangeEvent {
-  public target: IEventTarget;
-
-  constructor(name: string, value: any) {
-    this.target = { name, value };
-  }
-}
 
 /**
  * This is what we write to the .json files. We use "raw" subtypes here to
@@ -403,13 +387,10 @@ enum DeathMarkers {
   ALL = 'All',
 }
 
-enum Colors {
-  UNCOMMON = '#1eff00',
-  RARE = '#0070dd',
-  EPIC = '#a335ee',
-  LEGENDARY = '#ff8000',
-  ARTIFACT = '#e6cc80	',
-  BLIZZARD = '#00ccff',
+enum MarkerColors {
+  WIN = 'rgba(30, 255, 0, 1)',
+  LOSS = 'rgba(255, 0, 0, 1)',
+  ENCOUNTER = 'rgba(163, 53, 238, 1)',
 }
 
 type VideoMarker = {
@@ -439,7 +420,6 @@ export {
   FileInfo,
   FileFinderCallbackType,
   VideoQueueItem,
-  FakeChangeEvent,
   Metadata,
   RendererVideo,
   RendererVideoState,
@@ -467,7 +447,7 @@ export {
   ConfigStage,
   DeathMarkers,
   VideoMarker,
-  Colors,
+  MarkerColors,
   MicStatus,
   Crashes,
   CrashData,