+++ docs/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 ### Added
+- [Issue 463](https://github.com/aza547/wow-recorder/issues/463) - Added a video clipping feature, includes a reworked to the video player controls and timeline markers.
+
 ### Fixed
 - [Issue 459](https://github.com/aza547/wow-recorder/issues/459) - Fix a bug where we didn't respect the content type settings.